+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.12.0] (Unreleased)
 
+- chore: replace usages of `Pipeline#halted!` with `Pipeline#halt!`
 - feat: create `Ductwork::Pipeline#halt!` method
 - chore: extract `Ductwork::Pipeline#complete!` method